# refactor: Rework client timeout