--- a/src/apify_client/clients/base/resource_client.py
+++ b/src/apify_client/clients/base/resource_client.py
@@ -11,12 +11,13 @@ from apify_client.clients.base.base_client import BaseClient, BaseClientAsync
 class ResourceClient(BaseClient):
     """Base class for sub-clients manipulating a single resource."""
 
-    def _get(self) -> dict | None:
+    def _get(self, timeout_secs: int | None = None) -> dict | None:
         try:
             response = self.http_client.call(
                 url=self.url,
                 method='GET',
                 params=self._params(),
+                timeout_secs=timeout_secs,
             )
 
             return parse_date_fields(pluck_data(response.json()))
@@ -26,22 +27,24 @@ class ResourceClient(BaseClient):
 
         return None
 
-    def _update(self, updated_fields: dict) -> dict:
+    def _update(self, updated_fields: dict, timeout_secs: int | None = None) -> dict:
         response = self.http_client.call(
             url=self._url(),
             method='PUT',
             params=self._params(),
             json=updated_fields,
+            timeout_secs=timeout_secs,
         )
 
         return parse_date_fields(pluck_data(response.json()))
 
-    def _delete(self) -> None:
+    def _delete(self, timeout_secs: int | None = None) -> None:
         try:
             self.http_client.call(
                 url=self._url(),
                 method='DELETE',
                 params=self._params(),
+                timeout_secs=timeout_secs,
             )
 
         except ApifyApiError as exc:
@@ -52,12 +55,13 @@ class ResourceClient(BaseClient):
 class ResourceClientAsync(BaseClientAsync):
     """Base class for async sub-clients manipulating a single resource."""
 
-    async def _get(self) -> dict | None:
+    async def _get(self, timeout_secs: int | None = None) -> dict | None:
         try:
             response = await self.http_client.call(
                 url=self.url,
                 method='GET',
                 params=self._params(),
+                timeout_secs=timeout_secs,
             )
 
             return parse_date_fields(pluck_data(response.json()))
@@ -67,22 +71,24 @@ class ResourceClientAsync(BaseClientAsync):
 
         return None
 
-    async def _update(self, updated_fields: dict) -> dict:
+    async def _update(self, updated_fields: dict, timeout_secs: int | None = None) -> dict:
         response = await self.http_client.call(
             url=self._url(),
             method='PUT',
             params=self._params(),
             json=updated_fields,
+            timeout_secs=timeout_secs,
         )
 
         return parse_date_fields(pluck_data(response.json()))
 
-    async def _delete(self) -> None:
+    async def _delete(self, timeout_secs: int | None = None) -> None:
         try:
             await self.http_client.call(
                 url=self._url(),
                 method='DELETE',
                 params=self._params(),
+                timeout_secs=timeout_secs,
             )
 
         except ApifyApiError as exc: