@Pijukatel 2025-04-10
> I guess you don't use `timedelta` because there are already integers utilized, correct?

Yes, I used the same format that was already being used in the rest of the repository.