@Pijukatel 2025-04-22
Comparison of this change to master when used `BeautifulsoupCrawler`:

Without the change, the timeout issue caused 2 out of 10 crawler runs to be delayed for several minutes. On **average runtime was 110 s**
[Benchmark data](https://api.apify.com/v2/key-value-stores/AEuLwbSBF2QQ3hzpY/records/2025-04-22T06-47-18)
`
 2025-04-22 06:47:07,141 - benchmark_logger - INFO - Benchmark of run uHUeOuhjYGKTcfULc, benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160, Runtime: 55.079s, 
2025-04-22 06:47:08,457 - benchmark_logger - INFO - Benchmark of run DRYRXo0DwjCc9cxil, benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160, Runtime: 52.431s, 
2025-04-22 06:47:09,719 - benchmark_logger - INFO - Benchmark of run kZKfpEZrwKSIHcf1X, benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160, Runtime: 48.069s, 
2025-04-22 06:47:11,015 - benchmark_logger - INFO - Benchmark of run TldmdpnCzs7jZtbBX, benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160, Runtime: 364.947s, 
2025-04-22 06:47:12,337 - benchmark_logger - INFO - Benchmark of run SHSTq8RB83VS9eACy, benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160, Runtime: 48.108s, 
2025-04-22 06:47:13,703 - benchmark_logger - INFO - Benchmark of run 2t9bybamUDI5A0gvp, benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160, Runtime: 42.485s, 
2025-04-22 06:47:14,971 - benchmark_logger - INFO - Benchmark of run 3RmzOyDKrGeIt3phf, benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160, Runtime: 355.603s, 
2025-04-22 06:47:16,172 - benchmark_logger - INFO - Benchmark of run 2w84fA5giuVZtq3Ig, benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160, Runtime: 49.454s, 
2025-04-22 06:47:17,398 - benchmark_logger - INFO - Benchmark of run TgdXHgEu7VrWWdjxs, benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160, Runtime: 45.008s, 
2025-04-22 06:47:18,597 - benchmark_logger - INFO - Benchmark of run ZfUB4J2PFsi0eKIrq, benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160, Runtime: 44.959s, 
2025-04-22 06:47:18,598 - benchmark_logger - INFO - Overall benchmark of all runs, :aggregated_benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160.0, Runtime: 110.6143s,
`


With the change, the timeout issue caused 1 out of 10 crawler runs to be delayed for reduced timeout. On **average runtime was 47 s**
[Benchmark data](https://api.apify.com/v2/key-value-stores/AEuLwbSBF2QQ3hzpY/records/2025-04-22T10-33-12)
`
2025-04-22 10:33:04,939 - benchmark_logger - INFO - Benchmark of run O4sCXyUh4Bq123K4A, benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160, Runtime: 49.955s, 
2025-04-22 10:33:05,799 - benchmark_logger - INFO - Benchmark of run l6thld9FAQ3NZGYbU, benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160, Runtime: 46.766s, 
2025-04-22 10:33:06,666 - benchmark_logger - INFO - Benchmark of run 0fVrzImRHkmhlp1PU, benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160, Runtime: 44.235s, 
2025-04-22 10:33:07,448 - benchmark_logger - INFO - Benchmark of run wsKJuPOJwRh50q05I, benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160, Runtime: 42.443s, 
2025-04-22 10:33:08,338 - benchmark_logger - INFO - Benchmark of run 2lVzPb8qgVgv5b3h7, benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160, Runtime: 47.248s, 
2025-04-22 10:33:09,194 - benchmark_logger - INFO - Benchmark of run cTwxscer7IbmnrFJC, benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160, Runtime: 44.459s, 
2025-04-22 10:33:10,099 - benchmark_logger - INFO - Benchmark of run dtm1AvHdqQbKCtgrS, benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160, Runtime: 66.829s, 
2025-04-22 10:33:10,915 - benchmark_logger - INFO - Benchmark of run 8htGLQh32E4RdxeyN, benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160, Runtime: 42.451s, 
2025-04-22 10:33:11,789 - benchmark_logger - INFO - Benchmark of run ZCWnhLjHD6lcaarnm, benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160, Runtime: 43.601s, 
2025-04-22 10:33:12,564 - benchmark_logger - INFO - Benchmark of run Xo6dQd2zwkG3kOoQE, benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160, Runtime: 42.236s, 
2025-04-22 10:33:12,564 - benchmark_logger - INFO - Overall benchmark of all runs, :aggregated_benchmark=Actor: beautifulsoup-crawler-benchmark, Valid results: 160.0, Runtime: 47.0223s
`